Move tuivisualzer into its own cpp file
