+++ src/Visualizer/CMakeLists.txt
@@ -1,1 +1,1 @@
-ADD_LIBRARY(sir-visualizer STATIC visualizer.cpp)
+ADD_LIBRARY(sir-visualizer STATIC visualizer.cpp tui-visualizer.cpp)
